requirements.txt
numpy
opencv-python
matplotlib
python-dotenv
onvif-zeep